You know what's better than adding move semantics and initializer lists to the Array class?  Completely deleting all that code and using a std::vector directly :)
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@3054 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/gui/ComboDialog.cxx
@@ -114,7 +114,7 @@ void ComboDialog::loadConfig()
   StringList events;
   instance().eventHandler().getComboListForEvent(myComboEvent, events);
 
-  int size = BSPF_min(events.size(), 8u);
+  int size = BSPF_min((int)events.size(), 8);
   for(int i = 0; i < size; ++i)
     myEvents[i]->setSelected("", events[i]);
